Add a test for a ziggurat search bug.
When there are discontinuities in the pdf, there are multiple valid y's for each x. The ziggurat search algorithm needs to be aware of this. This example currently errors because when we move x from 8.0 to prevfloat(8.0) the y value and the layer area change discontinuously so the height of the ziggurat also changes discontinuously, and in this case it's impossible to make y[end] approximately equal pdf(0).