Rewrite to support discontinuous distributions
We needed a change of strategy in order to properly support discontinuous pdfs. Instead of searching for `x1`, (the width of the base not counting the tail) and building the ziggurat from there. We instead search for `y1` and let `x1 = ipdf(y1)`. This resolves a problem where  when changing x1 from one side of a discontinuity to another we skip over the optimal layer area since y jumped from one value to another. Instead we need to allow y to be any value between the discontinuous points. That way the layer area is a continuous function of the search variable and an optimal ziggurat can always be found.

At the same time we have simplified the implementation by putting x0, and y0=0 in the x and y vectors respectively as the first element (so the implementation ends up having x[1] = x0, everything is shifted up by one). Also bounded, unbounded, increasing, and decreasing cases share a lot more code now.