Fix a bug with single layer ziggurats.
This is a workaround to find the handedness of a ziggurat when x2 = modalboundary. I.e. when the pdf is strictly decreasing in a neighborhood of the modal boundary, so x2 = ipdf(pdf(modalboundary)) = modalboundary. It can still fail for distributions that are delta-like. I.e., ipdf(y2/2) = modalboundary. I might just have to ask the user not to supply functions that have isolated points? I need to do more testing to decide. This works for now on a broad class of functions.